Add experimental rewrite of git_batch
In it, we rely on git branches and many remotes for student management
instead of the old file based workflow.

This has the benefit of being able to keep commits better. The
implementation also appears to be simpler. Some operations are redundant
compared to the old git_batch. But in reality, this really should be
implemented in a non-string based programming language, probably.